+++ interfaces/src/main/java/ar/edu/itba/paw/persistance/RolesDao.java
@@ -14,28 +14,28 @@ public interface RolesDao {
     /**
      * Gets a user's role at a restaurant, or empty if said user has no roles at said restaurant.
      */
-    Optional<RestaurantRoleLevel> getRole(int userId, int restaurantId);
+    Optional<RestaurantRoleLevel> getRole(long userId, long restaurantId);
 
     /**
      * Sets a user's role at a restaurant. Specify roleLevel as null to remove a user's role at a restaurant.
      * @return True if the operation was successful
      */
-    boolean setRole(int userId, int restaurantId, RestaurantRoleLevel roleLevel);
+    boolean setRole(long userId, long restaurantId, RestaurantRoleLevel roleLevel);
 
     /**
      * Returns whether a given user has a given role level or higher at a given restaurant.
      */
-    boolean doesUserHaveRole(int userId, int restaurantId, RestaurantRoleLevel minimumRoleLevel);
+    boolean doesUserHaveRole(long userId, long restaurantId, RestaurantRoleLevel minimumRoleLevel);
 
     /**
      * Gets the users with roles for a restaurant, ordered by role in descending permissions.
      */
-    List<Pair<User, RestaurantRoleLevel>> getByRestaurant(int restaurantId);
+    List<Pair<User, RestaurantRoleLevel>> getByRestaurant(long restaurantId);
 
     /**
      * Gets the restaurants and roles for a given user, and the amount of non-finalized orders.
      */
-    List<Triplet<Restaurant, RestaurantRoleLevel, Integer>> getByUser(int userId);
+    List<Triplet<Restaurant, RestaurantRoleLevel, Integer>> getByUser(long userId);
 
-    boolean deleteRole(int restaurantId, int userId);
+    boolean deleteRole(long restaurantId, long userId);
 }
